Changes in VO and DTO for hibernate-VALIDATE.

+++ src/main/java/gmoldes/domain/dto/ClientDTO.java
@@ -13,9 +13,9 @@ public class ClientDTO {
     private Integer id;
     private Integer idcliente;
     private String nifcif;
-    private Integer nifcif_dup;
+    private Short nifcif_dup;
     private String nom_rzsoc;
-    private Integer numvez;
+    private Short numvez;
     private Character cltsg21;
     private Date fdesde;
     private Date fhasta;
@@ -26,8 +26,8 @@ public class ClientDTO {
     public ClientDTO() {
     }
 
-    public ClientDTO(Integer id, Integer idcliente, String nifcif, Integer nifcif_dup, String nom_rzsoc,
-                     Integer numvez, Character cltsg21, Date fdesde, Date fhasta, Boolean cltactivo,
+    public ClientDTO(Integer id, Integer idcliente, String nifcif, Short nifcif_dup, String nom_rzsoc,
+                     Short numvez, Character cltsg21, Date fdesde, Date fhasta, Boolean cltactivo,
                      Date sinactividad, String tipoclte) {
         this.id = id;
         this.idcliente = idcliente;
@@ -67,11 +67,11 @@ public class ClientDTO {
         this.nifcif = nifcif;
     }
 
-    public Integer getNifcif_dup() {
+    public Short getNifcif_dup() {
         return nifcif_dup;
     }
 
-    public void setNifcif_dup(Integer nifcif_dup) {
+    public void setNifcif_dup(Short nifcif_dup) {
         this.nifcif_dup = nifcif_dup;
     }
 
@@ -83,11 +83,11 @@ public class ClientDTO {
         this.nom_rzsoc = nom_rzsoc;
     }
 
-    public Integer getNumvez() {
+    public Short getNumvez() {
         return numvez;
     }
 
-    public void setNumvez(Integer numvez) {
+    public void setNumvez(Short numvez) {
         this.numvez = numvez;
     }
 
@@ -153,9 +153,9 @@ public class ClientDTO {
         private Integer id;
         private Integer idcliente;
         private String nifcif;
-        private Integer nifcif_dup;
+        private Short nifcif_dup;
         private String nom_rzsoc;
-        private Integer numvez;
+        private Short numvez;
         private Character cltsg21;
         private Date fdesde;
         private Date fhasta;
@@ -178,7 +178,7 @@ public class ClientDTO {
             return this;
         }
 
-        public ClientDTO.PersonBuilder withNifcif_dup(Integer nifcif_dup) {
+        public ClientDTO.PersonBuilder withNifcif_dup(Short nifcif_dup) {
             this.nifcif_dup = nifcif_dup;
             return this;
         }
@@ -188,7 +188,7 @@ public class ClientDTO {
             return this;
         }
 
-        public ClientDTO.PersonBuilder withNumvez(Integer numvez) {
+        public ClientDTO.PersonBuilder withNumvez(Short numvez) {
             this.numvez = numvez;
             return this;
         }
